Move v1 request tests to new folder
