+++ tests/requests/v1/favorites/create.spec.js
