+++ tests/requests/v1/favorites/show.spec.js
