+++ tests/requests/v1/playlistFavorites/create.spec.js
