+++ tests/requests/v1/playlistFavorites/show.spec.js
