+++ tests/requests/v1/playlists/create.spec.js
