+++ tests/requests/v1/playlists/destroy.spec.js
